+++ include/emanecontrolmessages/universalphycontrolsendmessage.h
@@ -371,7 +371,7 @@ namespace EMANE
 
 
       // scale factor for tx power
-      static const float DB_SCALE_FACTOR = 100.0;
+      static const int DB_SCALE_FACTOR = 100;
 
       // option list limits
       const static ACE_UINT8 MAX_NUM_FREQUENCY_INFO_ENTRIES = 255;
